requirements.txt
Flask>=3.0.3
click>=8.1.7
pytest>=8.3.3
pytest-flask>=1.3.0
python-dotenv>=1.0.1
PyYAML>=6.0